+++ monkey/monkey_island/cc/app.py
@@ -21,18 +21,18 @@ from monkey_island.cc.resources.netmap import NetMap
 from monkey_island.cc.resources.node import Node
 from monkey_island.cc.resources.remote_run import RemoteRun
 from monkey_island.cc.resources.reporting.report import Report
-from monkey_island.cc.resources.representations import output_json
 from monkey_island.cc.resources.root import Root
 from monkey_island.cc.resources.telemetry import Telemetry
 from monkey_island.cc.resources.telemetry_feed import TelemetryFeed
 from monkey_island.cc.resources.pba_file_download import PBAFileDownload
 from monkey_island.cc.resources.version_update import VersionUpdate
-from monkey_island.cc.services.database import Database
-from monkey_island.cc.consts import MONKEY_ISLAND_ABS_PATH
-from monkey_island.cc.services.remote_run_aws import RemoteRunAwsService
 from monkey_island.cc.resources.pba_file_upload import FileUpload
 from monkey_island.cc.resources.attack.attack_config import AttackConfiguration
 from monkey_island.cc.resources.attack.attack_report import AttackReport
+from monkey_island.cc.services.database import Database
+from monkey_island.cc.services.remote_run_aws import RemoteRunAwsService
+from monkey_island.cc.services.representations import output_json
+from monkey_island.cc.consts import MONKEY_ISLAND_ABS_PATH
 
 from monkey_island.cc.resources.test.monkey_test import MonkeyTest
 from monkey_island.cc.resources.test.log_test import LogTest
